+++ src/app/page.js
@@ -1,23 +1,20 @@
 import Countdown from "./ui/Countdown";
 import FaqSection from "./ui/FaqSection.jsx";
-
-async function getData() {
-  const res = await fetch("http://localhost:3001/sections");
-  if (!res.ok) {
-    throw new Error("Failed to fetch data");
-  }
-  return res.json();
-}
+import { getData } from "./service/FetchData";
+import Navbar from "./ui/Navbar";
 
 export default async function Home() {
   const data = await getData();
-  const countdown = data[0];
-  const { faqs } = data[0].faqSection;
+
+  const [countdownData] = data.sections;
+  const { navbar } = data;
+  const faqs = countdownData.faqSection.faqs
 
   return (
     <div>
+      <Navbar data={navbar} />
+      <Countdown data={countdownData.countdown} />
       <FaqSection faqs={faqs} />
-      <Countdown data={countdown} />
     </div>
   );
 }